linkedin-automator/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
websockets==13.1
google-genai>=1.66.0
supabase==2.9.0
python-dotenv==1.0.1
httpx==0.27.0
pydantic==2.9.0
pydantic-settings==2.5.0
pytest==8.3.0
pytest-asyncio==0.24.0
